// frontend/jobapptrackerfrontend/src/enums/editjobapptestids/EditJobAppTestids_enum.ts
export enum EditJobAppPageTestIds {
    header = "editJobAppPageHeader",
    jobTitle = "editJobAppPageJobTitle",
    jobTitleError = "editJobAppPageJobTitleHelper",
    companyError = "editJobAppPageCompanyError",
    companyField = "editJobAppPageCompany",
    descriptionField = "editJobAppPageDescription",
    statusField = "editJobAppPageStatus",
    statusFieldError = "editobAppPageStatusError",
    dateAppliedField = "editJobAppPageDateAppliedField",
    dateAppliedError = "editJobAppPageDateAppliedError",
    jobInterviewCard = "editJobAppPageInterviewCard_",
    oldDateApplied = "oldDateApplied",
    submitJobAppButton = "editJobAppPageSubmitJobApp",
    addInterviewButton = "addJobAppPageAddInterview"
};