// frontend/jobapptrackerfrontend/src/enums/jobinterviewtestids/AddInterviewModalTestIds_enum.ts
export enum AddInterviewModalTestIds {
    interviewType = "addInterviewTypeField",
    interviewTypeError = "addInterviewTypeError",
    locationField = "addInterviewLocationField",
    startDateField = "addInterviewStartDateField",
    startDateError = "addInterviewStartDateError",
    endDateField = "addInterviewEndDateField",
    endDateError = "addInterviewEndDateError",
    cancelButton = "addInterviewCancelButton",
    addInterviewButton = "addInterviewAddInterviewButton"
}